feat: userState module added to store
